compiling version of reformatting program
small and independent library+executable that translates
the Boost.Serialization archive output from hexactrl-sw
into the ROOT-based raw data format that will be input into ldmx-sw.

This translation layer also inserts headers that are expected to apear
in the final data pipeline allowing the ldmx-sw side to be closer to
what will actually be required for data taking.